pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "medagent"
version = "0.2.0"
description = "Agentic RAG MedMCQA: medBERT/sciBERT embeddings, FAISS retrieval, OpenAI Chat/Assistants, llm/datasette CLI, FastAPI."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Nik Jois"}]
dependencies = [
    "typer[all]>=0.12.3",
    "rich>=13.7.1",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "openai>=1.40.0",
    "tiktoken>=0.7.0",
    "transformers>=4.41.0",
    "torch>=2.2.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "sqlite-utils>=3.36",
    "datasette>=0.64.6",
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "jinja2>=3.1.4",
    "orjson>=3.10.0",
    "uvloop; platform_system == 'Linux'",
    "pyyaml>=6.0.1",
    "tenacity>=8.3.0",
    "tabulate>=0.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.6.9",
    "mypy>=1.11.0",
    "types-requests",
    "types-PyYAML",
]
cli = [
    "llm>=0.13",
]
data = [
    "datasets>=2.20.0",
]

[project.scripts]
medagent = "medagent.cli:app"
medagent-api = "medagent.api:run"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true

